requirements.txt
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
coveralls==1.8.1
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-nose==1.4.6
# djangorestframework==3.10.1
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
docopt==0.6.2
idna==2.8
nose==1.3.7
psycopg2-binary==2.8.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
